refactor(application-components): move modal-page-top-bar file now that we export it
