+++ packages/application-components/src/components/modal-pages/utils/index.ts
@@ -1,0 +1,4 @@
+export {
+  default as ModalPageTopBar,
+  LargeIconWrapper,
+} from './modal-page-top-bar';
